Moved native app artifacts logic to common API (#1877)
