+++ src/snowflake/cli/_plugins/nativeapp/codegen/setup/native_app_setup_processor.py
@@ -23,7 +23,6 @@ from typing import List, Optional
 import yaml
 from click import ClickException
 from snowflake.cli._plugins.nativeapp.artifacts import (
-    BundleMap,
     find_manifest_file,
     find_setup_script_file,
 )
@@ -37,6 +36,7 @@ from snowflake.cli._plugins.nativeapp.codegen.sandbox import (
     execute_script_in_sandbox,
 )
 from snowflake.cli._plugins.stage.diff import to_stage_path
+from snowflake.cli.api.artifacts.bundle_map import BundleMap
 from snowflake.cli.api.console import cli_console as cc
 from snowflake.cli.api.project.schemas.v1.native_app.path_mapping import (
     PathMapping,
