+++ src/snowflake/cli/_plugins/nativeapp/codegen/snowpark/python_processor.py
@@ -22,7 +22,6 @@ from typing import Any, Dict, List, Optional, Set
 
 from pydantic import ValidationError
 from snowflake.cli._plugins.nativeapp.artifacts import (
-    BundleMap,
     find_setup_script_file,
 )
 from snowflake.cli._plugins.nativeapp.codegen.artifact_processor import (
@@ -48,6 +47,7 @@ from snowflake.cli._plugins.nativeapp.codegen.snowpark.models import (
     NativeAppExtensionFunction,
 )
 from snowflake.cli._plugins.stage.diff import to_stage_path
+from snowflake.cli.api.artifacts.bundle_map import BundleMap
 from snowflake.cli.api.cli_global_context import get_cli_context, span
 from snowflake.cli.api.console import cli_console as cc
 from snowflake.cli.api.metrics import CLICounterField
